Refactor alternate dim, add new files to mental health
