+++ alternate_dimension/setup.py
@@ -6,6 +6,6 @@ setup(
    description='Predict alternate dimension transport',
    author='Michael Krug',
    author_email='michaelkrug92@gmail.com',
-   packages=['predict_transport'],
+   packages=['alternate_dimension'],
    install_requires=['pandas', 'numpy', 'scikit-learn', 'xgboost', 'pytest'],
 )
